fix: refatora logica de ParseApiError para pasta compartilhada
