+++ src/features/admin/api/parseApiError.ts
@@ -1,38 +1,9 @@
-type ParsedApiError = {
-  message: string;
-  fieldErrors: Record<string, string>;
-};
+import {
+  parseApiError as parseSharedApiError,
+  type ParsedApiError,
+} from '@/shared/parseApiError';
 
-export const parseApiError = (errorBody: unknown): ParsedApiError => {
-  if (!errorBody || typeof errorBody !== 'object') {
-    return {
-      message: 'Não foi possível cadastrar o usuário.',
-      fieldErrors: {},
-    };
-  }
+const DEFAULT_MESSAGE = 'Não foi possível cadastrar o usuário.';
 
-  const body = errorBody as {
-    message?: string;
-    errors?: Record<string, string[] | undefined>;
-  };
-
-  const fieldErrors: Record<string, string> = {};
-
-  if (body.errors) {
-    for (const [field, messages] of Object.entries(body.errors)) {
-      if (messages?.length) {
-        fieldErrors[field] = messages[0];
-      }
-    }
-  }
-
-  const firstFieldError = Object.values(fieldErrors)[0];
-
-  return {
-    message:
-      firstFieldError ||
-      body.message ||
-      'Não foi possível cadastrar o usuário.',
-    fieldErrors,
-  };
-};
+export const parseApiError = (errorBody: unknown): ParsedApiError =>
+  parseSharedApiError(errorBody, DEFAULT_MESSAGE);
